requirements_versions.txt
torch==2.4.0
torchvision~=0.19.0
torchsde==0.2.6
einops==0.4.1
transformers==4.44.0
safetensors==0.4.3
accelerate==0.33.0
pyyaml==6.0
Pillow==9.2.0
scipy==1.10.0
tqdm==4.64.1
psutil==5.9.5
pytorch_lightning==1.9.4
omegaconf==2.2.3
httpx==0.24.1
onnxruntime==1.16.3
timm==0.9.2
aiohttp~=3.8.5
SQLAlchemy[asyncio]~=2.0.19
asyncmy~=0.2.8
pydantic[dotenv]~=1.10.9
aiofiles~=23.2.1
aiosqlite~=0.19.0
opennsfw2~=0.13.7
tensorflow~=2.15.0
gradio==3.50.2
pygit2==1.12.2
diffusers==0.30.0
ftfy==6.1.1
clip_interrogator==0.6.0
opencv-python==4.8.1.78
kornia==0.7.2
sentencepiece==0.2.0
spandrel==0.3.4
imageio[pyav]==2.34.1 
xformers~=0.0.26
beautifulsoup4==4.12.3
rembg==2.0.58
trimesh==4.4.4
jaxtyping==0.2.33
gpytoolbox==0.3.2
pynim==1.0.5
slangtorch==1.2.3
numpy-stl==3.1.2
